+++ dropwizard-guicey/src/main/java/ru/vyarus/dropwizard/guice/debug/report/guice/util/visitor/GuiceScopingVisitor.java
@@ -1,5 +1,6 @@
 package ru.vyarus.dropwizard.guice.debug.report.guice.util.visitor;
 
+import com.google.inject.Binding;
 import com.google.inject.Scope;
 import com.google.inject.Scopes;
 import com.google.inject.Singleton;
@@ -7,7 +8,9 @@ import com.google.inject.servlet.RequestScoped;
 import com.google.inject.servlet.ServletScopes;
 import com.google.inject.servlet.SessionScoped;
 import com.google.inject.spi.DefaultBindingScopingVisitor;
+import com.google.inject.spi.LinkedKeyBinding;
 import ru.vyarus.dropwizard.guice.module.installer.feature.eager.EagerSingleton;
+import ru.vyarus.dropwizard.guice.module.installer.util.BindingUtils;
 import ru.vyarus.dropwizard.guice.module.support.scope.Prototype;
 
 import java.lang.annotation.Annotation;
@@ -18,6 +21,18 @@ import java.lang.annotation.Annotation;
  */
 public class GuiceScopingVisitor
         extends DefaultBindingScopingVisitor<Class<? extends Annotation>> {
+
+    /**
+     * Method to call DIRECTLY on visitor instead of "normal" visitor appliance. Required for more accurate
+     * scope resolution.
+     *
+     * @param binding binding to analyze
+     * @return resolved scope (or proposed prototype scope when scope not detected)
+     */
+    @SuppressWarnings("unchecked")
+    public Class<? extends Annotation> performDetection(final Binding binding) {
+        return finalizeScopeDetection((Class<? extends Annotation>) binding.acceptScopingVisitor(this), binding);
+    }
 
     @Override
     public Class<? extends Annotation> visitEagerSingleton() {
@@ -55,9 +70,31 @@ public class GuiceScopingVisitor
 
     @Override
     public Class<? extends Annotation> visitNoScoping() {
-        // special case: when checking direct module elements guice know only directly configured scope info and
-        // ignore annotations.. so instead of correct scope from annotation no scope is returned
+        // no scope annotation declared OR linked binding with scope annotation on TARGET class (no way to know it)
 
-        return Prototype.class;
+        return null;
+    }
+
+    /**
+     * Method should be called manually! Scoping visitor would not detect scoping annotation on linked binding.
+     * This method tries to fix this (to improve accuracy).
+     *
+     * @param scope   resolved scope or null
+     * @param binding binding under analysis
+     * @return scoping annotation (exactly resolved or assumed prototype)
+     */
+    private Class<? extends Annotation> finalizeScopeDetection(final Class<? extends Annotation> scope,
+                                                               final Binding binding) {
+        if (scope != null) {
+            return scope;
+        }
+        Class<? extends Annotation> res = null;
+        if (binding instanceof LinkedKeyBinding) {
+            // NOTE: the link may be a part of long chain, but this is ignored - consider only length of 1
+            // (which may obviously produce false scope value)
+            final LinkedKeyBinding linkedBinding = (LinkedKeyBinding) binding;
+            res = BindingUtils.findScopingAnnotation(linkedBinding.getLinkedKey().getTypeLiteral().getRawType(), true);
+        }
+        return res == null ? Prototype.class : visitScopeAnnotation(res);
     }
 }