+++ dropwizard-guicey/src/main/java/ru/vyarus/dropwizard/guice/module/installer/feature/jersey/AbstractJerseyInstaller.java
@@ -1,18 +1,15 @@
 package ru.vyarus.dropwizard.guice.module.installer.feature.jersey;
 
 import com.google.inject.Binder;
-import com.google.inject.ScopeAnnotation;
 import com.google.inject.binder.AnnotatedBindingBuilder;
+import jakarta.inject.Singleton;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import ru.vyarus.dropwizard.guice.module.installer.FeatureInstaller;
 import ru.vyarus.dropwizard.guice.module.installer.install.JerseyInstaller;
 import ru.vyarus.dropwizard.guice.module.installer.install.binding.LazyBinding;
 import ru.vyarus.dropwizard.guice.module.installer.option.InstallerOptionsSupport;
-
-import jakarta.inject.Scope;
-import jakarta.inject.Singleton;
-import java.lang.annotation.Annotation;
+import ru.vyarus.dropwizard.guice.module.installer.util.BindingUtils;
 
 import static ru.vyarus.dropwizard.guice.module.installer.InstallersOptions.ForceSingletonForJerseyExtensions;
 import static ru.vyarus.dropwizard.guice.module.installer.InstallersOptions.JerseyExtensionsManagedByGuice;
@@ -92,19 +89,6 @@ public abstract class AbstractJerseyInstaller<T> extends InstallerOptionsSupport
      * @return true if scope annotation found, false otherwise
      */
     private boolean hasScopeAnnotation(final Class<?> type, final boolean hkManaged) {
-        boolean found = false;
-        for (Annotation ann : type.getAnnotations()) {
-            final Class<? extends Annotation> annType = ann.annotationType();
-            if (annType.isAnnotationPresent(Scope.class)) {
-                found = true;
-                break;
-            }
-            // guice has special marker annotation
-            if (!hkManaged && annType.isAnnotationPresent(ScopeAnnotation.class)) {
-                found = true;
-                break;
-            }
-        }
-        return found;
+        return BindingUtils.findScopingAnnotation(type, !hkManaged) != null;
     }
 }