+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 * Update to dropwizard 5 (requires java 17)
+* Improve guice bindings report:
+    - fixed scope accuracy for linked bindings
 * [admin-rest]
   - Add identifyAdminContextInRequestLogs bundle option to highlight admin requests in logs